services/chatbot-service/requirements.txt
Django==5.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
requests==2.32.3
chromadb==0.5.23
sentence-transformers==3.4.1
google-generativeai==0.8.4
